+++ app/parse-siti-payment-file/value.js
@@ -1,7 +1,7 @@
 const { convertToPence } = require('../currency-convert')
 
-const getInvoiceTotalInPence = (data, key) => {
+const getValueInPence = (data, key) => {
   return data.reduce((a, b) => { return a + convertToPence(b[key]) }, 0)
 }
 
-module.exports = getInvoiceTotalInPence
+module.exports = getValueInPence
